+++ CromIAM/src/main/scala/cromiam/instrumentation/CromIamInstrumentation.scala
@@ -2,13 +2,12 @@ package cromiam.instrumentation
 
 import akka.http.scaladsl.model.{HttpRequest, HttpResponse}
 import cats.data.NonEmptyList
+import cromwell.api.model.FailureResponseOrT
 import cromwell.core.instrumentation.InstrumentationPrefixes._
 import cromwell.services.instrumentation.CromwellInstrumentation
 import cromwell.services.instrumentation.CromwellInstrumentation.InstrumentationPath
 
-import scala.concurrent.{ExecutionContext, Future}
 import scala.concurrent.duration._
-import scala.util.{Failure, Success}
 
 object CromIamInstrumentation {
   private val UUIDRegex = {
@@ -49,17 +48,28 @@ trait CromIamInstrumentation extends CromwellInstrumentation {
 
   def instrumentationPrefixForSam(methodPrefix: NonEmptyList[String]): NonEmptyList[String] = samPrefix.concatNel(methodPrefix)
 
-  def instrumentRequest[A](func: () => Future[A], httpRequest: HttpRequest, prefix: NonEmptyList[String])(implicit ec: ExecutionContext): Future[A] = {
-    val startTimestamp = System.currentTimeMillis
+  def instrumentRequest[A](func: () => FailureResponseOrT[A],
+                           httpRequest: HttpRequest,
+                           prefix: NonEmptyList[String]): FailureResponseOrT[A] = {
+    def now(): Deadline = Deadline.now
 
-    val funcResponseFuture = func()
+    val startTimestamp = now()
 
-    funcResponseFuture.onComplete {
-      case Success(response: HttpResponse) => sendTimingApi(makePathFromRequestAndResponse(httpRequest, response), (System.currentTimeMillis - startTimestamp).millis, prefix)
-      case Success(_) => sendTimingApi(makePathFromRequestAndResponseString(httpRequest, "success"), (System.currentTimeMillis - startTimestamp).millis, prefix)
-      case Failure(_) => sendTimingApi(makePathFromRequestAndResponseString(httpRequest, "failure"), (System.currentTimeMillis - startTimestamp).millis, prefix)
+    def sendDuration(instrumentationPath: InstrumentationPath): Unit = {
+      val duration = now() - startTimestamp
+      sendTimingApi(instrumentationPath, duration, prefix)
     }
 
-    funcResponseFuture
+    func() map {
+      case response: HttpResponse =>
+        sendDuration(makePathFromRequestAndResponse(httpRequest, response))
+        response.asInstanceOf[A]
+      case other =>
+        sendDuration(makePathFromRequestAndResponseString(httpRequest, "success"))
+        other.asInstanceOf[A]
+    } leftMap { error =>
+      sendDuration(makePathFromRequestAndResponseString(httpRequest, "failure"))
+      error
+    }
   }
 }