+++ CromIAM/src/test/scala/cromiam/webservice/SwaggerUiHttpServiceSpec.scala
@@ -1,14 +1,14 @@
 package cromiam.webservice
 
 import akka.http.scaladsl.model.headers.Location
-import akka.http.scaladsl.model.{StatusCodes, Uri}
+import akka.http.scaladsl.model.{ContentTypes, StatusCodes, Uri}
 import akka.http.scaladsl.server.Route
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 import com.typesafe.config.ConfigFactory
 import cromiam.server.config.SwaggerOauthConfig
+import cromiam.webservice.SwaggerUiHttpServiceSpec._
 import org.scalatest.prop.TableDrivenPropertyChecks
 import org.scalatest.{FlatSpec, Matchers}
-import cromiam.webservice.SwaggerUiHttpServiceSpec._
 
 trait SwaggerUiHttpServiceSpec extends FlatSpec with Matchers with ScalatestRouteTest with SwaggerUiHttpService {
   def actorRefFactory = system
@@ -46,12 +46,14 @@ class BasicSwaggerUiHttpServiceSpec extends SwaggerUiHttpServiceSpec {
     Get() ~> swaggerUiRoute ~> check {
       status should be(StatusCodes.TemporaryRedirect)
       header("Location") should be(Option(Location(Uri("/swagger"))))
+      contentType should be(ContentTypes.`text/html(UTF-8)`)
     }
   }
 
   it should "not return options for /" in {
     Options() ~> Route.seal(swaggerUiRoute) ~> check {
       status should be(StatusCodes.MethodNotAllowed)
+      contentType should be(ContentTypes.`text/plain(UTF-8)`)
     }
   }
 
@@ -59,6 +61,7 @@ class BasicSwaggerUiHttpServiceSpec extends SwaggerUiHttpServiceSpec {
     Get("/swagger") ~> swaggerUiRoute ~> check {
       status should be(StatusCodes.TemporaryRedirect)
       header("Location") should be(Option(Location(Uri("/swagger/index.html?url=/api-docs"))))
+      contentType should be(ContentTypes.`text/html(UTF-8)`)
     }
   }
 
@@ -66,6 +69,7 @@ class BasicSwaggerUiHttpServiceSpec extends SwaggerUiHttpServiceSpec {
     Get("/swagger/index.html") ~> swaggerUiRoute ~> check {
       status should be(StatusCodes.OK)
       responseAs[String] should include("replaced-client-id")
+      contentType should be(ContentTypes.`text/html(UTF-8)`)
     }
   }
 }
@@ -78,12 +82,14 @@ class NoRedirectRootSwaggerUiHttpServiceSpec extends SwaggerUiHttpServiceSpec {
   it should "not redirect / to /swagger" in {
     Get() ~> Route.seal(swaggerUiRoute) ~> check {
       status should be(StatusCodes.NotFound)
+      contentType should be(ContentTypes.`text/plain(UTF-8)`)
     }
   }
 
   it should "not return options for /" in {
     Options() ~> Route.seal(swaggerUiRoute) ~> check {
       status should be(StatusCodes.MethodNotAllowed)
+      contentType should be(ContentTypes.`text/plain(UTF-8)`)
     }
   }
 
@@ -91,12 +97,14 @@ class NoRedirectRootSwaggerUiHttpServiceSpec extends SwaggerUiHttpServiceSpec {
     Get("/swagger") ~> swaggerUiRoute ~> check {
       status should be(StatusCodes.TemporaryRedirect)
       header("Location") should be(Option(Location(Uri("/swagger/index.html?url=/api-docs"))))
+      contentType should be(ContentTypes.`text/html(UTF-8)`)
     }
   }
 
   it should "return index.html from the swagger-ui jar" in {
     Get("/swagger/index.html") ~> swaggerUiRoute ~> check {
       status should be(StatusCodes.OK)
+      contentType should be(ContentTypes.`text/html(UTF-8)`)
     }
   }
 }
@@ -110,6 +118,7 @@ class DefaultSwaggerUiConfigHttpServiceSpec extends SwaggerUiHttpServiceSpec wit
     Get("/swagger") ~> swaggerUiRoute ~> check {
       status should be(StatusCodes.TemporaryRedirect)
       header("Location") should be(Option(Location(Uri("/swagger/index.html?url=/api-docs"))))
+      contentType should be(ContentTypes.`text/html(UTF-8)`)
     }
   }
 
@@ -117,6 +126,7 @@ class DefaultSwaggerUiConfigHttpServiceSpec extends SwaggerUiHttpServiceSpec wit
     Get("/swagger/index.html") ~> swaggerUiRoute ~> check {
       status should be(StatusCodes.OK)
       responseAs[String].take(SwaggerIndexPreamble.length) should be(SwaggerIndexPreamble)
+      contentType should be(ContentTypes.`text/html(UTF-8)`)
     }
   }
 }
@@ -136,6 +146,7 @@ class OverriddenSwaggerUiConfigHttpServiceSpec extends SwaggerUiHttpServiceSpec 
     Get("/ui/path") ~> swaggerUiRoute ~> check {
       status should be(StatusCodes.TemporaryRedirect)
       header("Location") should be(Option(Location(Uri("/base/ui/path/index.html?url=/base/swagger/cromiam.yaml"))))
+      contentType should be(ContentTypes.`text/html(UTF-8)`)
     }
   }
 
@@ -143,6 +154,7 @@ class OverriddenSwaggerUiConfigHttpServiceSpec extends SwaggerUiHttpServiceSpec 
     Get("/ui/path/index.html") ~> swaggerUiRoute ~> check {
       status should be(StatusCodes.OK)
       responseAs[String].take(SwaggerIndexPreamble.length) should be(SwaggerIndexPreamble)
+      contentType should be(ContentTypes.`text/html(UTF-8)`)
     }
   }
 }
@@ -156,18 +168,21 @@ class YamlSwaggerResourceHttpServiceSpec extends SwaggerResourceHttpServiceSpec 
     Get("/swagger/testservice.yaml") ~> swaggerResourceRoute ~> check {
       status should be(StatusCodes.OK)
       responseAs[String] should startWith("swagger: '2.0'\n")
+      contentType should be(ContentTypes.`application/octet-stream`)
     }
   }
 
   it should "not service swagger json" in {
     Get("/swagger/testservice.json") ~> Route.seal(swaggerResourceRoute) ~> check {
       status should be(StatusCodes.NotFound)
+      contentType should be(ContentTypes.`text/plain(UTF-8)`)
     }
   }
 
   it should "not service /swagger" in {
     Get("/swagger") ~> Route.seal(swaggerResourceRoute) ~> check {
       status should be(StatusCodes.NotFound)
+      contentType should be(ContentTypes.`text/plain(UTF-8)`)
     }
   }
 
@@ -176,6 +191,7 @@ class YamlSwaggerResourceHttpServiceSpec extends SwaggerResourceHttpServiceSpec 
       Options(path) ~> swaggerResourceRoute ~> check {
         status should be(StatusCodes.OK)
         responseAs[String] should be("OK")
+        contentType should be(ContentTypes.`text/plain(UTF-8)`)
       }
     }
   }
@@ -192,18 +208,21 @@ class JsonSwaggerResourceHttpServiceSpec extends SwaggerResourceHttpServiceSpec 
     Get("/swagger/testservice.json") ~> swaggerResourceRoute ~> check {
       status should be(StatusCodes.OK)
       responseAs[String] should startWith("{\n  \"swagger\": \"2.0\",\n")
+      contentType should be(ContentTypes.`application/json`)
     }
   }
 
   it should "not service swagger yaml" in {
     Get("/swagger/testservice.yaml") ~> Route.seal(swaggerResourceRoute) ~> check {
       status should be(StatusCodes.NotFound)
+      contentType should be(ContentTypes.`text/plain(UTF-8)`)
     }
   }
 
   it should "not service /swagger" in {
     Get("/swagger") ~> Route.seal(swaggerResourceRoute) ~> check {
       status should be(StatusCodes.NotFound)
+      contentType should be(ContentTypes.`text/plain(UTF-8)`)
     }
   }
 
@@ -212,6 +231,7 @@ class JsonSwaggerResourceHttpServiceSpec extends SwaggerResourceHttpServiceSpec 
       Options(path) ~> swaggerResourceRoute ~> check {
         status should be(StatusCodes.OK)
         responseAs[String] should be("OK")
+        contentType should be(ContentTypes.`text/plain(UTF-8)`)
       }
     }
   }
@@ -228,18 +248,21 @@ class NoOptionsSwaggerResourceHttpServiceSpec extends SwaggerResourceHttpService
     Get("/swagger/testservice.yaml") ~> swaggerResourceRoute ~> check {
       status should be(StatusCodes.OK)
       responseAs[String] should startWith("swagger: '2.0'\n")
+      contentType should be(ContentTypes.`application/octet-stream`)
     }
   }
 
   it should "not service swagger json" in {
     Get("/swagger/testservice.json") ~> Route.seal(swaggerResourceRoute) ~> check {
       status should be(StatusCodes.NotFound)
+      contentType should be(ContentTypes.`text/plain(UTF-8)`)
     }
   }
 
   it should "not service /swagger" in {
     Get("/swagger") ~> Route.seal(swaggerResourceRoute) ~> check {
       status should be(StatusCodes.NotFound)
+      contentType should be(ContentTypes.`text/plain(UTF-8)`)
     }
   }
 
@@ -247,6 +270,7 @@ class NoOptionsSwaggerResourceHttpServiceSpec extends SwaggerResourceHttpService
     forAll(testPathsForOptions) { path =>
       Options(path) ~> Route.seal(swaggerResourceRoute) ~> check {
         status should be(StatusCodes.MethodNotAllowed)
+        contentType should be(ContentTypes.`text/plain(UTF-8)`)
       }
     }
   }
@@ -263,6 +287,7 @@ class YamlSwaggerUiResourceHttpServiceSpec extends SwaggerUiResourceHttpServiceS
     Get("/swagger") ~> swaggerUiResourceRoute ~> check {
       status should be(StatusCodes.TemporaryRedirect)
       header("Location") should be(Option(Location(Uri("/swagger/index.html?url=/swagger/testservice.yaml"))))
+      contentType should be(ContentTypes.`text/html(UTF-8)`)
     }
   }
 
@@ -270,12 +295,14 @@ class YamlSwaggerUiResourceHttpServiceSpec extends SwaggerUiResourceHttpServiceS
     Get("/swagger/testservice.yaml") ~> swaggerUiResourceRoute ~> check {
       status should be(StatusCodes.OK)
       responseAs[String] should startWith("swagger: '2.0'\n")
+      contentType should be(ContentTypes.`application/octet-stream`)
     }
   }
 
   it should "not service swagger json" in {
     Get("/swagger/testservice.json") ~> Route.seal(swaggerUiResourceRoute) ~> check {
       status should be(StatusCodes.NotFound)
+      contentType should be(ContentTypes.`text/plain(UTF-8)`)
     }
   }
 
@@ -284,6 +311,7 @@ class YamlSwaggerUiResourceHttpServiceSpec extends SwaggerUiResourceHttpServiceS
       Options(path) ~> swaggerUiResourceRoute ~> check {
         status should be(StatusCodes.OK)
         responseAs[String] should be("OK")
+        contentType should be(ContentTypes.`text/plain(UTF-8)`)
       }
     }
   }
@@ -303,6 +331,7 @@ class JsonSwaggerUiResourceHttpServiceSpec extends SwaggerUiResourceHttpServiceS
     Get("/swagger") ~> swaggerUiResourceRoute ~> check {
       status should be(StatusCodes.TemporaryRedirect)
       header("Location") should be(Option(Location(Uri("/swagger/index.html?url=/swagger/testservice.json"))))
+      contentType should be(ContentTypes.`text/html(UTF-8)`)
     }
   }
 
@@ -310,12 +339,14 @@ class JsonSwaggerUiResourceHttpServiceSpec extends SwaggerUiResourceHttpServiceS
     Get("/swagger/testservice.json") ~> swaggerUiResourceRoute ~> check {
       status should be(StatusCodes.OK)
       responseAs[String] should startWith("{\n  \"swagger\": \"2.0\",\n")
+      contentType should be(ContentTypes.`application/json`)
     }
   }
 
   it should "not service swagger yaml" in {
     Get("/swagger/testservice.yaml") ~> Route.seal(swaggerUiResourceRoute) ~> check {
       status should be(StatusCodes.NotFound)
+      contentType should be(ContentTypes.`text/plain(UTF-8)`)
     }
   }
 
@@ -324,6 +355,7 @@ class JsonSwaggerUiResourceHttpServiceSpec extends SwaggerUiResourceHttpServiceS
       Options(path) ~> swaggerUiResourceRoute ~> check {
         status should be(StatusCodes.OK)
         responseAs[String] should be("OK")
+        contentType should be(ContentTypes.`text/plain(UTF-8)`)
       }
     }
   }